Changing styles and changed css files to scss. Pt 1
Modified the style files to 1 work now that they are in smaller module scss files instead of an overall css one, and 2 changed a bit of the appearance for the styles. The styles for the footer and majority of the navbar (except adjustments based on screen size) have been updated. Pt 2 would be updating the events and home page module scss files to work with the jsx and seeing if the same can be done for the responsive screen adjustments