Merging working frontend aspect to main for progress purposes (#25)
* footer created

* footer-embedded map and links

* Events Page files created

* calendar added

* events grid and list partially completed

* Past Events Added

* Past Events Updated

* changed to module.css

* Added assets and reorganized styles

As discussed in a meeting with the frontend, the styles were split into smaller module css files. The next step is to update the jsx files to use the classes from the module css imports

* Changing styles and changed css files to scss. Pt 1

Modified the style files to 1 work now that they are in smaller module scss files instead of an overall css one, and 2 changed a bit of the appearance for the styles. The styles for the footer and majority of the navbar (except adjustments based on screen size) have been updated. Pt 2 would be updating the events and home page module scss files to work with the jsx and seeing if the same can be done for the responsive screen adjustments

* Changed a lot Pt 2

Changed styles (both colour, default elements, and some created styles), organizational structures, and modified the default template

---------

Co-authored-by: Toff-cowan <cowanchristoff24@gmail.com>